# docs/contracts/api.openapi.yaml
openapi: 3.1.0
info:
  title: VideoCutlist API
  version: 1.0.0
  description: >-
    Deployment authentication applies to application endpoints. API responses carry
    a server-generated X-Request-ID; errors contain error.code, error.message, and
    matching error.requestId. Missing/invalid authentication is 401, rejected origins
    are 403 origin_forbidden, and unexpected persistence/runtime failures are safe
    500 responses. Existing 422 input-validation exceptions are retained. HEAD
    responses have no body. MCP JSON-RPC is a separate protocol outside this contract.
servers:
  - url: "{serverBaseUrl}/api/v1"
    variables:
      serverBaseUrl:
        default: http://localhost:8787
        description: >-
          Reachable VideoCutlist server base URL; transport provider
          independent.
security:
  - bearerAuth: []
  - {}
paths:
...
  /media:
    get:
      summary: List indexed media
      operationId: listMedia
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Media page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaPage"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
...
  /media/refresh:
    post:
      summary: Refresh the media index
      operationId: refreshMedia
      responses:
        "202":
          description: Durable library scan job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportJob"
        "403":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /media/{mediaId}:
    parameters:
      - $ref: "#/components/parameters/MediaId"
    get:
      summary: Read media metadata
      operationId: getMedia
      responses:
        "200":
          description: Media metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"
        "404":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
...
  /projects:
    get:
      summary: List saved projects
      operationId: listProjects
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Project page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectPage"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      summary: Read a project
      operationId: getProject
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
    put:
      summary: Save a project
      operationId: putProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "200":
          description: Saved project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /projects/{projectId}/detections:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    post:
      summary: Start media detection
      operationId: createDetection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DetectionInput"
      responses:
        "202":
          description: Detection queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DetectionJob"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /projects/{projectId}/exports:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    post:
      summary: Create a durable export
      operationId: createExport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportInput"
      responses:
        "202":
          description: Batch export queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchExportSubmission"
        "404":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /batches:
    get:
      summary: List durable batches
      operationId: listBatches
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Recent durable batches
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchPage"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /batches/{batchId}:
    parameters:
      - name: batchId
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get durable batch export progress
      operationId: getBatchProgress
      responses:
        "200":
          description: Batch progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
        "404":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
    delete:
      summary: Cancel a batch export
      operationId: cancelBatch
      responses:
        "204":
          description: >-
            Cancellation is requested for each queued or running child; 204 does
            not mean every runner has exited or its resources are released.
            Terminal children remain unchanged.
        "404":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
...
  /projects/{projectId}/exports/preflight:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    post:
      summary: Validate export selection without queueing
      operationId: preflightExport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportInput"
      responses:
        "200":
          description: Structured export findings and deterministic selection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportPreflight"
        "422":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
...
  /jobs/{jobId}/retry:
    parameters:
      - $ref: "#/components/parameters/JobId"
    post:
      summary: Retry a failed export as a new job
      operationId: retryJob
      responses:
        "202":
          description: New retry batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Used only when the server selects static bearer mode.
  parameters:
...
    MediaId:
      name: mediaId
      in: path
      required: true
      schema:
        type: string
        pattern: ^m_[A-Za-z0-9_-]{43}$
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        pattern: ^p_[A-Za-z0-9_-]{12,64}$
    JobId:
      name: jobId
      in: path
      required: true
      schema:
        type: string
        pattern: ^j_[A-Za-z0-9_-]{12,64}$
  responses:
    Error:
      description: Safe error; the request ID matches X-Request-ID
      headers:
        X-Request-ID: {schema: {type: string}}
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    ImportJob:
      type: object
      required: [id, state, progress, indexed]
      properties:
        id: {type: string, pattern: '^j_[A-Za-z0-9_-]{12,64}$'}
        state: {type: string, enum: [queued, running, succeeded, failed, cancelled]}
        progress: {type: number, minimum: 0, maximum: 1}
        indexed: {type: integer, minimum: 0}
        rootResults:
          type: object
          additionalProperties:
            type: object
            required: [state]
            properties:
              state: {type: string, enum: [scanning, ready_empty, ready_with_media, failed]}
              errorCode: {type: string, enum: [scan_limit, cancelled, scan_failed]}
        errorCode: {type: string}
        validationErrors:
          type: array
          items: {type: string}
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - requestId
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
    Media:
      type: object
      required:
        - id
        - name
        - durationMs
        - sizeBytes
        - container
        - streams
        - etag
      properties:
        id:
          type: string
        name:
          type: string
        durationMs:
          type: integer
        sizeBytes:
          type: integer
        container:
          type: string
        streams:
          type: object
          additionalProperties: true
        etag:
          type: string
...
    MediaPage:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Media"
        nextCursor:
          type:
            - string
            - "null"
...
    ProjectInput:
      type: object
      additionalProperties: false
      required:
        - revision
        - schemaVersion
        - name
        - items
      properties:
        revision:
          type: integer
          minimum: 0
        schemaVersion:
          type: integer
          const: 2
        name:
          type: string
          minLength: 1
          maxLength: 120
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ProjectItem"
    ProjectSummary:
      type: object
      required:
        - id
        - name
        - revision
        - updatedAt
      properties:
        id:
          type: string
        name:
          type: string
        revision:
          type: integer
          minimum: 1
        updatedAt:
          type: string
          format: date-time
    ProjectPage:
      type: object
      required:
        - items
        - nextCursor
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ProjectSummary"
        nextCursor:
          type: [string, "null"]
    Project:
      allOf:
        - $ref: "#/components/schemas/ProjectInput"
        - type: object
          required:
            - id
            - updatedAt
          properties:
            id:
              type: string
            updatedAt:
              type: string
              format: date-time
    ProjectItem:
      type: object
      additionalProperties: false
      required:
        - id
        - mediaId
        - segments
        - exportOptions
      properties:
        id:
          type: string
          pattern: ^i_[A-Za-z0-9_-]{24}$
        mediaId:
          type: string
          pattern: ^m_[A-Za-z0-9_-]{43}$
        segments:
          type: array
          items:
            $ref: "#/components/schemas/Segment"
        editorState:
          $ref: "#/components/schemas/UIState"
        exportOptions:
          $ref: "#/components/schemas/ProjectExportOptions"
    ProjectExportOptions:
      type: object
      additionalProperties: false
      properties:
        mode:
          type: string
          enum: [merge, separate]
        selection:
          type: string
          enum: [segments, gaps]
        streamIndexes:
          type: array
          uniqueItems: true
          items:
            type: integer
            minimum: 0
        cutStrategy:
          type: string
          enum: [stream_copy_preferred, precise_reencode, hybrid_smart_cut]
        container:
          type: string
          enum: [mkv, mp4, mov]
        destinationId:
          type: string
          maxLength: 64
        filenameTemplate:
          type: string
          maxLength: 160
    Segment:
      type: object
      additionalProperties: false
      required:
        - startMs
        - endMs
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 128
        startMs:
          type: integer
          minimum: 0
        endMs:
          type: integer
          minimum: 1
        label:
          type: string
          maxLength: 200
        included:
          type: boolean
          default: true
    UIState:
      type: object
      required:
        - playheadMs
        - zoom
        - muted
      properties:
        playheadMs:
          type: integer
          minimum: 0
        zoom:
          type: number
          exclusiveMinimum: 0
        muted:
          type: boolean
    DetectionInput:
      type: object
      required: [mediaId, projectItemId, projectRevision, kind]
      properties:
        mediaId:
          type: string
        projectItemId:
          type: string
        projectRevision:
          type: integer
          minimum: 1
        kind:
          type: string
          enum: [silence, black, scene]
        sourceFingerprint:
          type: string
        noiseDb:
          type: number
        minDurationMs:
          type: integer
          minimum: 0
        sceneThreshold:
          type: number
          minimum: 0
          maximum: 1
    DetectionCandidate:
      oneOf:
        - $ref: "#/components/schemas/DetectionRangeCandidate"
        - $ref: "#/components/schemas/DetectionPointCandidate"
    DetectionRangeCandidate:
      type: object
      additionalProperties: false
      required: [id, mediaId, projectId, projectRevision, startMs, endMs, source]
      properties:
        id: {type: string}
        mediaId: {type: string}
        projectId: {type: string}
        projectRevision: {type: integer}
        startMs: {type: integer, minimum: 0}
        endMs: {type: integer, minimum: 1}
        source: {type: string, enum: [silence, black]}
...
    DetectionJob:
      type: object
      required: [id, type, state, mediaId, projectId, projectRevision, kind]
      properties:
        id: {type: string}
        type: {type: string, enum: [detection]}
        state:
          type: string
          enum: [queued, running, succeeded, failed, cancelled]
        mediaId: {type: string}
        projectId: {type: string}
        projectRevision: {type: integer}
        kind: {type: string, enum: [silence, black, scene]}
        candidates:
          type: array
          items: {$ref: "#/components/schemas/DetectionCandidate"}
        errorCode: {type: string}
    ExportInput:
      type: object
      additionalProperties: false
      required:
        - mode
        - selection
        - cutStrategy
        - container
      properties:
        mode:
          type: string
          enum:
            - merge
            - separate
        selection:
          type: string
          enum:
            - segments
            - gaps
          default: segments
        streamIndexes:
          type: array
          uniqueItems: true
          items:
            type: integer
            minimum: 0
        cutStrategy:
          type: string
          enum:
            - stream_copy_preferred
            - precise_reencode
            - hybrid_smart_cut
        container:
          type: string
          enum:
            - mkv
            - mp4
            - mov
        destinationId:
          type: string
          maxLength: 64
          description: >-
            Opaque configured destination ID; never a filesystem path.
        filenameTemplate:
          type: string
          maxLength: 160
          description: Restricted template using documented variables only.
        itemIds:
          type: array
          uniqueItems: true
          items:
            type: string
          description: >-
            Optional ordered project-item selection; omitted selects all.
    BatchExportSubmission:
      type: object
      required:
        - batchId
        - jobs
      properties:
        batchId:
          type: string
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
    Batch:
      type: object
      required:
        - batchId
        - state
        - progress
        - jobs
      properties:
        batchId:
          type: string
        projectId:
          type: string
        projectRevision:
          type: integer
          minimum: 1
        state:
          type: string
          enum: [queued, running, succeeded, failed, cancelled]
        progress:
          type: number
          minimum: 0
          maximum: 1
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
    BatchPage:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Batch"
...
    ExportPreflight:
      type: object
      required:
        - allowed
        - selection
        - findings
      properties:
        allowed:
          type: boolean
        selection:
          type: array
          items:
            type: integer
            minimum: 0
        findings:
          type: array
          items:
            $ref: "#/components/schemas/ExportFinding"
    ExportFinding:
      type: object
      required:
        - severity
        - code
        - message
      properties:
        severity:
          type: string
          enum:
            - allowed
            - warn
            - blocked
        code:
          type: string
        message:
          type: string
        streamIndex:
          type: integer
          minimum: 0
...
    Job:
      type: object
      required:
        - id
        - type
        - state
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        batchId:
          type: string
        projectItemId:
          type: string
        mediaLabel:
          type: string
        mediaId:
          type: string
        projectId:
          type: string
        projectRevision:
          type: integer
        kind:
          type: string
        type:
          type: string
          enum:
            - export
            - detection
            - library_scan
        state:
          type: string
          enum:
            - queued
            - running
            - succeeded
            - failed
            - cancelled
        progress:
          type: number
          minimum: 0
          maximum: 1
        result:
          description: Present only when an export job succeeded.
          type: object
          additionalProperties: false
          required:
            - sizeBytes
            - retainUntil
          properties:
            container:
              type: string
              enum: [mkv, mp4, mov]
              description: Published output container.
            outputName:
              type: string
              description: >-
                Published merged-export filename, never a filesystem path.
            outputNames:
              type: array
              description: >-
                Published separate-export filenames, never filesystem paths.
              items:
                type: string
            outputFailures:
              type: array
              description: >-
                Per-segment failures when a separate export partially succeeds.
              items:
                type: object
                additionalProperties: false
                required: [segment, code, message]
                properties:
                  segment:
                    type: integer
                    minimum: 1
                  code:
                    type: string
                  message:
                    type: string
            sizeBytes:
              type: integer
              format: int64
              minimum: 0
            retainUntil:
              type: string
              format: date-time
            destinationId:
              type: string
            destinationKind:
              type: string
              enum:
                - download
                - archive
                - source_adjacent
            appliedStrategies:
              type: array
              description: >-
                Strategy used for each selected segment; outputName identifies
                separately published outputs.
              items:
                type: object
                additionalProperties: false
                required:
                  - segment
                  - strategy
                properties:
                  segment:
                    type: integer
                    minimum: 1
                  outputName:
                    type: string
                    description: >-
                      Published separate-export filename, never a filesystem
                      path.
                  strategy:
                    type: string
        strategy:
          type: string
          description: Requested cut strategy.
        container:
          type: string
          enum: [mkv, mp4, mov]
          description: Requested output container.
        mode:
          type: string
          enum: [merge, separate]
          description: Requested output arrangement.
        selection:
          type: string
          enum: [segments, gaps]
          description: Requested ranges (included cuts or gaps between cuts).
        selectedStreams:
          type: array
          items:
            type: integer
            minimum: 0
          description: Resolved stream indexes used by the export job.
        appliedStrategy:
          type: string
          description: >-
            Strategy used by every selected segment; omitted when segments
            use different strategies. See result.appliedStrategies for
            per-segment values.
        warnings:
          type: array
          items:
            type: string
        warningDetails:
          type: array
          items:
            type: object
            required:
              - severity
              - code
              - message
            properties:
              severity:
                type: string
              code:
                type: string
              message:
                type: string
              streamIndex:
                type: integer
                minimum: 0
        verified:
          type: boolean
          description: False when the output requires human inspection.
        errorCode:
          type: string
          description: >-
            Stable export failure code; present only when state is failed.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time